Remove Game class, major refactor
Only unit tests are working.  Integration and E2E still failing.

It may be necessary to rewrite the (young, small) repository and API at
this point, since both were based around the primary concept of saving
and loading Games, which is no longer supported.

The new concept of the service is to offer only the calculation
features, with any state tracking being minimal and secondary.  Hence
the new centerpiece domain service is `calculate_deductions`.